Fixes #15
This bug was created with the addition of font variations after refactoring some of the core - Commit ID: 216b054046d3ee4664585047109394ce304ebe2e

When AdobeBlank fonts were used, the SizeWatcher was configured to watch for a decrease of the watched element size, instead of an increase. Elements with AdobeBlank fonts always have a width of 0, so when normal fonts are loaded the element grows in size.